Task2-Teeth classification using pre-trained model/teeth_classification_app/requirements.txt
streamlit==1.24.0
tensorflow==2.13.0
Pillow==9.5.0
numpy==1.23.5
pandas==2.0.3